refactor(MongoDB): renaming file to "MongoDB" instead of "MongoDb"
